refactor(#933): extract release-channel command runtime crate
